Seedify the Hanabi environment for purity of fitness function. Make fitness in whole architecture Double instead of Int.

+++ src/main/kotlin/ru/ifmo/genome/gamestrategies/core/Environment.kt
@@ -4,5 +4,5 @@ interface Environment<T : Individual<T>> {
     /**
      * Calculates fitness of single individual
      */
-    fun fit(individual : T) : Int
+    fun fit(individual : T) : Double
 }
